+++ frontend/src/pages/Home/Home.jsx
@@ -4,7 +4,6 @@ import "./Style.scss";
 import logoMS from "../../assets/pills.svg";
 import logoMS2 from "../../assets/pills2.svg";
 import logoMS3 from "../../assets/pills3.svg";
-import "./Style.scss";
 
 export default function Home() {
   return (
