+++ frontend/src/pages/Register/style.scss
@@ -38,7 +38,9 @@
     height: auto;
   }
 }
-
+.formclass {
+  margin-top: 5vh;
+}
 .avatarLogo {
   border-radius: 50%;
   box-shadow: 0 20px 60px 0 black;
@@ -92,8 +94,21 @@
 .favatar {
   display: flex;
   padding: 0rem;
-  margin: 0rem;
   padding-left: 0.8em;
+  max-height: 11em;
+  max-width: 11em;
+}
+.textregis {
+  display: flex;
+  justify-content: center;
+  color: whitesmoke;
+  text-shadow: 2px 2px black;
+  font-size: 1.9em;
+  font-weight: bolder;
+  text-decoration: underline 2px whitesmoke;
+  text-decoration-thickness: 3px;
+  margin-top: 2rem;
+  text-underline-offset: 5px;
 }
 .avatarPrev {
   display: flex;
@@ -102,6 +117,10 @@
   height: 10%;
   border-radius: 50%;
   margin-left: 44%;
+  max-height: 11em;
+  max-width: 11m;
+  margin-top: 7vh;
+  margin-bottom: 0vh;
 }
 
 .favatarInput,
@@ -119,15 +138,16 @@
 
 .validation {
   padding: 0.5em;
-  border-radius: 1em;
+  border-radius: 3em;
   font-family: App.$font-text;
   box-shadow: 0 20px 60px 0 black;
-  border: none;
   font-size: 1.2em;
   display: block;
   margin: auto;
   margin-top: 3%;
   color: App.$text-color-primary;
+  width: 7vw;
+  height: 7vh;
 }
 
 .validation {
@@ -136,6 +156,7 @@
 }
 
 .validation:hover {
+  cursor: pointer;
   background-color: #27ae60;
 }
 
@@ -149,20 +170,37 @@
 
 .logoMS2 {
   position: absolute;
-  height: 20em;
-  right: 5%;
-  top: 40%;
+  height: 14em;
+  right: -6%;
+  top: 85%;
   z-index: -5;
   transform: rotate(-120deg);
 }
 .logoMS3 {
+  margin-top: -1em;
+  position: absolute;
+  height: 14em;
+  transform: rotate(-200deg);
+  right: -6%;
+  top: -5%;
+}
+
+.logoMS4 {
+  position: absolute;
+  height: 11em;
+  right: 88%;
+  top: -4%;
+  z-index: -5;
+  transform: rotate(-110deg);
+}
+.logoMS5 {
   margin-top: 3em;
   position: absolute;
-  height: 7em;
-  transform: rotate(-220deg);
-  right: -2%;
-  top: -1%;
+  height: 15em;
+  transform: rotate(-10deg);
+  right: 88%;
+  top: 75%;
 }
+
 .subscriptionForm {
-  margin-top: 10em;
 }
